Minishell FINISHED

+++ src/execute/exec_redir.c
@@ -6,7 +6,7 @@
 /*   By: vhovhann <vhovhann@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/17 15:41:54 by vhovhann          #+#    #+#             */
-/*   Updated: 2023/09/10 21:14:31 by vhovhann         ###   ########.fr       */
+/*   Updated: 2023/09/12 16:48:47 by vhovhann         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,14 +33,14 @@ int	redir(t_main *main, t_tok *stack, t_env **env)
 		return (1);
 	}
 	tmp = stack;
-	while (tmp->left->type != WORD)
+	while (tmp->left->type != WORD && check_types(tmp->left->type) != 1)
 		tmp = tmp->left;
 	tmp->left->stdout_backup = main->stdout_backup;
 	tmp->left->_stdout_ = fd;
 	if (stack->last_red != 1)
 		return (0);
 	if (ft_strcmp(tmp->left->cmd, "(NULL)") && !(tmp->flag & (1 << 7)))
-		stack->err_code = cmds_execute(main, tmp->left, env, 0);
+		stack->err_code = check_astree(main, tmp->left, *env);
 	return (stack->err_code);
 }
 
@@ -58,13 +58,13 @@ int	heredoc(t_main *main, t_tok *stack, t_env **env)
 	tmp = stack;
 	while (tmp->left->type != WORD)
 	{
-		if (tmp->left->type == HEREDOC)
+		if (tmp->left->type == HEREDOC && check_types(tmp->left->type) != 1)
 			unlink(tmp->left->hdoc_fname);
 		tmp = tmp->left;
 	}
 	tmp->left->stdin_backup = main->stdin_backup;
 	tmp->left->_stdin_ = fd;
-	if (stack->last_hdoc != 1)
+	if (stack->last_hdoc != 1 || tmp->left->type == PIPE)
 		return (0 + unlink(stack->hdoc_fname));
 	if (ft_strcmp(tmp->left->cmd, "(NULL)"))
 		stack->err_code = check_astree(main, tmp->left, *env);
@@ -87,13 +87,13 @@ int	input(t_main *main, t_tok *stack, t_env **env)
 	if (i == 1)
 		return (1);
 	tmp = stack;
-	while (tmp->left->type != WORD)
+	while (tmp->left->type != WORD && check_types(tmp->left->type) != 1)
 		tmp = tmp->left;
 	tmp->left->stdin_backup = main->stdin_backup;
 	tmp->left->_stdin_ = fd;
 	if (stack->last_input != 1)
 		return (0);
-	stack->err_code = cmds_execute(main, tmp->left, env, 0);
+	stack->err_code = check_astree(main, tmp->left, *env);
 	i = 0;
 	return (0);
 }
@@ -101,8 +101,6 @@ int	input(t_main *main, t_tok *stack, t_env **env)
 int	exec_iocmd(t_main *main, t_tok *stack, t_env **env)
 {
 	check_lasts(main, stack, 0);
-	if (stack->left->left)
-		stack->err_code = check_astree(main, stack->left, *env);
 	if (main->exit_status == EXIT_SUCCESS)
 	{
 		if (stack && stack->type == HEREDOC)
